removed unnecearry includes, renamed files in mock_build from *dictionary.* to *mock.*
