+++ src/tests/mock_build/mock.ino
@@ -1,13 +1,9 @@
 #include <Arduino.h>
-#include <SPI.h>
-#include <SD.h>
-#include "test_dictionary.h"
+#include "test_mock.h"
 
 void
 setup(
 ) {
-	SPI.begin();
-	SD.begin(SD_CS_PIN);
 	Serial.begin(BAUD_RATE);
 	runalltests_mock();
 }
