+++ src/tests/mock_build/run_mock.c
@@ -1,4 +1,4 @@
-#include "test_dictionary.h"
+#include "test_mock.h"
 
 int
 main(
