+++ src/tests/mock_build/test_mock.h
@@ -5,10 +5,6 @@
 #include <stdio.h>
 #include <string.h>
 #include "../planckunit/src/planck_unit.h"
-#include "../../dictionary/dictionary_types.h"
-#include "../../dictionary/dictionary.h"
-#include "../../dictionary/ion_master_table.h"
-#include "../../dictionary/flat_file/flat_file_dictionary_handler.h"
 
 #ifdef  __cplusplus
 extern "C" {
